refactor: 패키지 구조 변경 및 Output 클래스 추가 적용
