+++ src/main/java/racingcar/view/RacingCarOutput.java
@@ -1,0 +1,15 @@
+package racingcar.view;
+
+public class RacingCarOutput {
+
+    private RacingCarOutput() {
+    }
+    public static final String SYSTEM_MESSAGE_FINAL_WINNER = "최종 우승자: ";
+    public static final String ERROR_CAR_NAME_LENGTH_MAX_FIVE = "자동차 이름의 길이가 5를 초과했습니다!";
+    public static final String ERROR_MESSAGE = "[ERROR] ";
+    public static final String ERROR_REFLECTION = "리플렉션 중 에러 발생";
+    public static final String ERROR_RANDOM_NUMBER_MAX_OVERFLOW = "끝값이 너무 큽니다. (스택 오버플로우 발생이 가능합니다)";
+    public static final String ERROR_RANDOM_NUMBER_MAX_LIMIT = "입력값이 너무 큽니다.";
+    public static final String ERROR_START_NOT_BIG_END = "startInclusive가 endInclusive보다 클 수 없습니다.";
+
+}
